Add files via upload
A script which evaluates the performance of a random forest validation set while iterating over preset values of the various parameters associated with a random forest, which includes n_estimators, max_depth, and min_samples_split. The performance was graphed with respect to the area under the ROC curve (a representation of the accuracy of the ROC Curve).